remove log-based incr backup tests/examples
