+++ examples/c/CMakeLists.txt
@@ -25,14 +25,6 @@ define_c_test(
 )
 
 define_c_test(
-    TARGET ex_backup_block
-    SOURCES ex_backup_block.c
-    DIR_NAME ex_backup_block
-    ARGUMENTS -h $<SHELL_PATH:$<TARGET_FILE_DIR:ex_backup_block>/WT_HOME>
-    DEPENDS "WT_POSIX"
-)
-
-define_c_test(
     TARGET ex_call_center
     SOURCES ex_call_center.c
     DIR_NAME ex_call_center
