+++ src/docs/examples.dox
@@ -3,10 +3,7 @@
 Create, insert and access a simple table.
 
 @example ex_backup.c
-Shows how to use log-based incremental backup.
-
-@example ex_backup_block.c
-Shows how to use block-based incremental backup.
+Shows how to use incremental backup.
 
 @example ex_call_center.c
 A more complex schema based on a call center example, showing how to map
