+++ test/evergreen/code_coverage/code_coverage_config.json
@@ -146,7 +146,6 @@
     "test/csuite/wt9937_parse_opts/test_wt9937_parse_opts",
     "test/csuite/wt3363_checkpoint_op_races/test_wt3363_checkpoint_op_races",
     "test/csuite/wt3120_filesys/test_wt3120_filesys -b $(pwd)",
-    "test/csuite/wt8659_reconstruct_database_from_logs/test_wt8659_reconstruct_database_from_logs",
 
     "python3 ../test/suite/run.py cursor_bound -s 1",
     "python3 ../test/suite/run.py cursor_bound_fuzz",
